+++ R/mock_boxplot.R
@@ -5,8 +5,8 @@
 #' @param ui_defaults,srv_defaults a list of values passed to the ui/server function
 #' @export
 
-mock_app_boxplot <- function(dry_run = FALSE, update_query_string = TRUE, srv_defaults = list(), ui_defaults = list()) {
-  data <- test_data()
+mock_app_boxplot <- function(dry_run = FALSE, update_query_string = TRUE, srv_defaults = list(), ui_defaults = list(), anlfl_flags = FALSE) {
+  data <- test_data(anlfl_flags = anlfl_flags)
   bm_dataset <- shiny::reactive({
     data[["bm"]]
   })
@@ -22,6 +22,12 @@ mock_app_boxplot <- function(dry_run = FALSE, update_query_string = TRUE, srv_de
     ui_defaults
   )
 
+  if (anlfl_flags) {
+    anlfl_vars <- c("ANLFL1", "ANLFL2")
+  } else {
+    anlfl_vars <- NULL
+  }
+
   srv_params <- c(
     list(
       id = "not_ebas",
@@ -31,7 +37,8 @@ mock_app_boxplot <- function(dry_run = FALSE, update_query_string = TRUE, srv_de
       cat_var = "PARCAT",
       par_var = "PARAM",
       visit_var = "VISIT",
-      value_vars = c("VALUE1", "VALUE2", "VALUE3")
+      value_vars = c("VALUE1", "VALUE2", "VALUE3"),
+      anlfl_vars = anlfl_vars
     ),
     srv_defaults
   )
@@ -49,18 +56,28 @@ mock_app_boxplot <- function(dry_run = FALSE, update_query_string = TRUE, srv_de
   )
 }
 
+
 #' Mock mm boxplot app
 #' @keywords mock
 #' @inheritParams mock_app_boxplot
 #' @export
 
-mock_app_boxplot_mm <- function(update_query_string = TRUE) {
+mock_app_boxplot_mm <- function(update_query_string = TRUE, anlfl_flags = FALSE) {
   if (!requireNamespace("dv.manager")) {
     stop("Install dv.manager")
   }
 
+  data <- test_data(anlfl_flags = anlfl_flags)
+
+
+  if (anlfl_flags) {
+    anlfl_vars <- c("ANLFL1", "ANLFL2")
+  } else {
+    anlfl_vars <- NULL
+  }
+
   dv.manager::run_app(
-    data = list(dummy = list(bm = test_data()[["bm"]], adsl = test_data()[["sl"]])),
+    data = list(dummy = list(bm = data[["bm"]], adsl = data[["sl"]])),
     module_list = list(
       Boxplot = mod_boxplot(
         "boxplot",
@@ -69,7 +86,8 @@ mock_app_boxplot_mm <- function(update_query_string = TRUE) {
         visit_var = "VISIT",
         value_vars = c("VALUE1", "VALUE2"),
         subjid_var = "SUBJID",
-        cat_var = "PARCAT"
+        cat_var = "PARCAT",
+        anlfl_vars = anlfl_vars
       )
     ),
     filter_data = "adsl",
@@ -77,3 +95,8 @@ mock_app_boxplot_mm <- function(update_query_string = TRUE) {
     enableBookmarking = "url"
   )
 }
+
+
+
+
+