+++ tests/testthat/test-scatterplotmatrix_app.R
@@ -11,7 +11,8 @@ ID <- poc( # nolint
     PAR = tns(SPM$ID$PAR, "par_val"),
     VAL = tns(SPM$ID$PAR_VALUE, "val"),
     VIS = tns(SPM$ID$PAR_VISIT, "val"),
-    GRP = tns(SPM$ID$MAIN_GRP, "val")
+    GRP = tns(SPM$ID$MAIN_GRP, "val"),
+    ANLFL = tns(BP$ID$ANLFL_FILTER, "val")
   ),
   OUTPUT = poc(
     CHART = tns(SPM$ID$CHART)
@@ -159,5 +160,41 @@ test_that("default values are set", {
   expect_equal(input_values[[ID$INPUT$VIS]], srv_defaults[["default_visit"]])
   expect_equal(input_values[[ID$INPUT$VAL]], srv_defaults[["default_value"]])
   expect_equal(input_values[[ID$INPUT$GRP]], srv_defaults[["default_main_group"]])
+})
+
+
+
+
+test_that("default values are set including analysis flag variables", {
+  testthat::skip_if_not(run_shiny_tests)
+  fail_if_app_not_started()
+  skip_if_suspect_check()
+
+  srv_defaults <- list(
+    default_cat = "PARCAT2",
+    default_par = c("PARAM22", "PARAM23"),
+    default_visit = "VISIT2",
+    default_value = "VALUE2",
+    default_main_group = "CAT3"
+  )
+
+  app <- start_app_driver(
+    rlang::quo(
+      dv.explorer.parameter::mock_app_scatterplotmatrix(
+        srv_defaults = !!srv_defaults,
+        anlfl_flags = TRUE
+      )
+    )
+  )
+
+  app$wait_for_idle()
+
+  input_values <- app$get_values()[["input"]]
+  expect_equal(input_values[[ID$INPUT$CAT]], srv_defaults[["default_cat"]])
+  expect_equal(input_values[[ID$INPUT$PAR]], srv_defaults[["default_par"]])
+  expect_equal(input_values[[ID$INPUT$VIS]], srv_defaults[["default_visit"]])
+  expect_equal(input_values[[ID$INPUT$VAL]], srv_defaults[["default_value"]])
+  expect_equal(input_values[[ID$INPUT$GRP]], srv_defaults[["default_main_group"]])
+  expect_equal(input_values[[ID$INPUT$ANLFL]], "ANLFL1")
 })
 # nolint end