+++ NEWS.md
@@ -1,3 +1,8 @@
+# dv.explorer.parameter 0.1.8-9000
+
+* boxplot, lineplot, scatter plot, scatter plot matrix, correlation heatmap, waterfall plus heatmap :
+    * Added the option to filter the input dataset based on Yes/No flag variables that exist in it.
+
 # dv.explorer.parameter 0.1.8
 
 * lineplot: